Refactoring periodic_log_writer
